% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{testshlib}
\alias{testshlib}
\title{Test system / R CMD SHLIB.}
\usage{
testshlib(dllname = "xyz")
}
\arguments{
\item{dllname}{character}
}
\description{
Test system / R CMD SHLIB.
}
\examples{
testshlib()

}

